Add useAuthorization custom hook
